feat: automatically detect the part of the solved riddle
* Added a feature to guess the part of the riddle that is
  being submitted by the user. The guessing is based on
  the description of the submitted riddle.
* Added a short contribution guide.
* Added a small `Help` section to the readme.

Closes #32
