+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "rust-analyzer.linkedProjects": [
+    "./Cargo.toml"
+  ]
+}
