+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Contribution
+
+Contributions are welcome!
+
+## Help
+
+If something is not working as expected, or you have an idea for an improvement,
+feel free to contact me at: `konrad.pagacz@gmail.com`, open an issue or a pull
+request.
+
+## Some conventions
+
+### Visibility and privacy of structures
+
+1. The default visibility of structures is `private`. Export structures as public only if needed.
+
+1. Only reexport structures at the top level of the crate for easier access for users.
