+++ Cargo.toml
@@ -2,7 +2,7 @@
 [package]
 name = "elv"
 description = "A little CLI helper for Advent of Code. 🎄"
-version = "0.12.3"
+version = "0.13.0"
 authors = ["Konrad Pagacz <konrad.pagacz@gmail.com>"]
 edition = "2021"
 readme = "README.md"
@@ -11,7 +11,6 @@ homepage = "https://github.com/kpagacz/elv"
 repository = "https://github.com/kpagacz/elv"
 keywords = ["adventofcode", "aoc", "cli", "cache", "rust"]
 categories = ["command-line-utilities"]
-
 
 [[bin]]
 name = "elv"
