+++ src/infrastructure/aoc_api.rs
@@ -1,4 +1,4 @@
-use crate::Configuration;
+use super::configuration::Configuration;
 
 const AOC_URL: &str = "https://adventofcode.com";
 
@@ -10,4 +10,5 @@ pub struct AocApi {
 
 mod aoc_api_impl;
 pub mod aoc_client_impl;
+pub mod find_riddle_part_impl;
 pub mod get_leaderboard_impl;
