+++ src/infrastructure/find_riddle_part.rs
@@ -1,0 +1,5 @@
+use crate::domain::RiddlePart;
+
+pub trait FindRiddlePart {
+    fn find(&self, year: i32, day: i32) -> Result<RiddlePart, anyhow::Error>;
+}
