+++ src/lib.rs
@@ -1,10 +1,7 @@
-pub mod application;
+mod application;
 pub mod domain;
-mod driver;
 mod infrastructure;
-mod submission_history;
 
-pub use crate::application::cli::CliCommand;
-pub use crate::application::cli::CliInterface;
-pub use crate::driver::Driver;
-pub use crate::infrastructure::Configuration;
+pub use crate::application::cli::ElvCli;
+pub use crate::infrastructure::configuration::Configuration;
+pub use crate::infrastructure::driver::Driver;
